requirements.txt
Django==2.2.1
djangorestframework==3.9.3
mysql-connector-python==8.0.13
protobuf==3.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0